Moved files, added comments + formatted files with black
